\name{gbt-package}
\alias{gbt-package}
\alias{gbtpackage}
\docType{package}
\title{
Gradient Boosted Trees
}
\description{
This package implements J. Friedman's gradient boosting machine.
It provides a parallel implementation of gradient boosting
with trees for regression problems (squared loss)
and binary classification problems (binomial deviance).
Includes built-in stratified cross-validation and
relative importance (relative influence) of features.
}
\details{
\tabular{ll}{
Package: \tab gbt\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-28\cr
License: \tab Something between AstraZeneca and The University of Manchester\cr
}
}
\author{
Alexandre Michelis <alexandremichelis@gmail.com>
}
\references{
J.H. Friedman, T. Hastie, R. Tibshirani (2009). \dQuote{The Elements of Statistical Learning: Data Mining, Inference, and
Prediction}.
}
