\name{process.features}
\alias{process.features}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Pre-processing of the features used for building a \code{gbt} object.
}
\description{
Helper function to build the regression trees.
}
\usage{
process.features(data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
The matrix to process.
}
}
\details{
This is not intended for end-user use.
}
\value{
Returned the pre-processed features.
}
\author{
Alexandre Michelis \email{alexandremichelis@gmail.com}
}
